一時保管。
