+++ ServiceApplication1/pch.h
@@ -8,6 +8,7 @@
 #include <Windows.h>
 #include <tchar.h>
 #include <winsvc.h>
+#include <processthreadsapi.h>
 
 #include <fstream>
